requirements.txt
flask==3.0.3
supabase==2.8.0          # latest stable as of 2026
python-dotenv==1.0.1
flask-cors==5.0.0        # allow frontend requests
gunicorn==23.0.0         # for production later